Official Beta Version of the Game
This was the submitted beta. Before animation, explosions. My goal for
next submission is to fix collisions by adding in the hit variable to
the lasers, asteroids, and ships to help limit collisions to a single
occurance, rather then the multiple same-time collisions.

Also would like to make a new "pong" like version, but that is heady and
may not happen.